// src/modules/branch/dto/store-branch.dto.ts
export interface StoreBranchDto {
  id?: number;
  entityId?: number;
  countryId?: number;
  matrix?: string;
  matrixId?: number;
  name?: string;
  chainName?: string;
  type?: number;
  address?: string;
  department?: string;
  city?: string;
  location?: string;
  phone?: string;
  whatsapp?: string;
  email?: string;
  hours?: string;
  countEmployees?: number;
  monthlySales?: number;
  localCurrency?: string;
  services?: string;
  licenseNumber?: string;
  localStatus?: string;
  chainType?: number;
  institutional?: string;
  specialized?: string;
  withinRetail?: string;
  salesCategory?: number;
  notes?: string;
  validationState?: number;
  trustLevel?: string;
  confirmedAddress?: string;
  validatedLicense?: string;
  confirmedLocation?: string;
  validatedBy?: number;
  validationDate?: Date;
  pointSaleState?: string;
  pointSalePlan?: string;
  pointSaleStartDate?: Date;
  pointSaleEndDate?: Date;
  pointSaleAutoRenewal?: string;
  users: { name: string } | null;
}